+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Additions
 
+- ChoiceBox now supports Commanding
 - TextField now supports Commanding
 - TreeTableSmartResize.POLICY - activate with smartResize() (https://github.com/edvin/tornadofx/issues/316)
 - removeWhen/visibleWhen/enableWhen/disableWhen etc functions now also take an observable instead of a function that returns an observable.